fix(modalcomposite): isPrompt adds mod-prompt css class
